// dashboard/src/utils/hasura-api/generated/schemas/addComputedFieldArgs.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { ComputedFieldDefinition } from './computedFieldDefinition';
import type { QualifiedTable } from './qualifiedTable';

export interface AddComputedFieldArgs {
  table: QualifiedTable;
  /** Name of the new computed field */
  name: string;
  /** The computed field definition */
  definition: ComputedFieldDefinition;
  /** Customize the description shown in GraphQL introspection. If null or omitted then if a comment exists on the database function, it is used as the description, and if not, an autogenerated description is used instead. */
  comment?: string;
  /** Name of the source database of the table */
  source?: string;
}